amdtelecom/requirements.txt
asgiref==3.3.1
Django==3.1.4
django-3-jet==1.0.8
django-colorfield==0.4.1
django-colorful==1.3
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-filter==2.4.0
django-phonenumber-field==5.0.0
django-resized==0.3.11
djangorestframework==3.12.2
phonenumbers==8.12.18
Pillow==8.0.1
psycopg2-binary==2.8.6
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
Unidecode==1.2.0
